+++ src/server/GameState.ts
@@ -1,7 +1,6 @@
 import { List, Map } from 'immutable';
-import { Observable } from 'rxjs';
-import { map, mergeAll, scan } from 'rxjs/operators';
-import { GameEvents, IncomingEvent, OutgoingEvent } from '../shared/events';
+import { OutgoingEvent } from '../shared/events';
+import { OutgoingEvents } from './events/OutgoingEvents';
 
 export class Player {
   public readonly id: number;
@@ -27,8 +26,7 @@ export class GameState {
   public newPlayer(id: number, socketId: string): GameState {
     this.players = this.players.set(id, (new Player(id, socketId)));
 
-    const newPlayer = new GameEvents.NewPlayer(id, socketId);
-    this.updates = this.updates.push(newPlayer, newPlayer.broadcast());
+    this.updates = this.updates.push(OutgoingEvents.NewPlayer.create(id, socketId));
 
     return this;
   }
@@ -36,7 +34,7 @@ export class GameState {
   public setPlayerName(id: number, name: string): GameState {
     this.players.get(id).name = name;
 
-    const updatePlayerInfo = new GameEvents.UpdatePlayerInfo(id, name);
+    const updatePlayerInfo = new OutgoingEvents.UpdatePlayerInfo(id, name);
     this.updates = this.updates.push(updatePlayerInfo);
 
     return this;
@@ -49,18 +47,3 @@ export class GameState {
   }
 }
 
-const accumulateGameState = (
-  gameState: GameState,
-  event: IncomingEvent,
-): GameState => {
-  return event.applyTo(gameState);
-};
-
-export const applyToGameState = () => (
-  source$: Observable<IncomingEvent>,
-): Observable<OutgoingEvent> =>
-  source$.pipe(
-    scan<IncomingEvent, GameState>(accumulateGameState, new GameState()),
-    map((gameState: GameState) => gameState.changes()),
-    mergeAll(),
-  );
