+++ src/server/ProcessIncomingEvents.ts
@@ -1,0 +1,20 @@
+import { Observable } from 'rxjs';
+import { map, mergeAll, scan } from 'rxjs/operators';
+import { IncomingEvent, OutgoingEvent } from '../shared/events';
+import { GameState } from './GameState';
+
+const applyToGameState = (
+  gameState: GameState,
+  event: IncomingEvent,
+): GameState => {
+  return event.applyTo(gameState);
+};
+
+export const processIncomingEvents = (initialState = new GameState()) => (
+  source$: Observable<IncomingEvent>,
+): Observable<OutgoingEvent> =>
+  source$.pipe(
+    scan<IncomingEvent, GameState>(applyToGameState, initialState),
+    map((gameState: GameState) => gameState.changes()),
+    mergeAll(),
+  );
