+++ src/shared/operators.ts
@@ -1,6 +1,6 @@
-import { Observable, pipe } from 'rxjs';
-import { filter, first, map, withLatestFrom } from 'rxjs/operators';
-import { BlobEvent, IncomingEvent } from './events';
+import { Observable, of, pipe } from 'rxjs';
+import { concatMap, filter, map, takeWhile, withLatestFrom } from 'rxjs/operators';
+import { IncomingEvent } from './events';
 
 export const unless = (locked$: Observable<any>) => (
   source$: Observable<any>,
@@ -11,14 +11,15 @@ export const unless = (locked$: Observable<any>) => (
     map(([event, _]) => event),
   );
 
-export const when = (enabled: boolean) => (toggle$: Observable<any>) =>
-  toggle$.pipe(filter((value) => value === enabled));
-
 export const eventsMatching = (type: string) =>
   pipe(filter((event: IncomingEvent) => event.type === type));
 
-export const firstMatching = <T extends BlobEvent>(type: string) => (source$: Observable<T>): Observable<T> =>
+export const allUntil = (type: string) => (
+  source$: Observable<IncomingEvent>,
+): Observable<IncomingEvent> =>
   source$.pipe(
-    filter((event: T) => event.type === type),
-    first(),
+    concatMap(
+      (next: IncomingEvent) => (next.type === type ? of(next, null) : of(next)),
+    ),
+    takeWhile<IncomingEvent>((next: IncomingEvent) => next !== null),
   );
